chore: move cypress folder into app/frontend
