+++ action/get-latest-visual-regression-status.ts
@@ -1,6 +1,6 @@
 import { octokit } from './octokit';
 import { context } from '@actions/github';
-import { VISUAL_REGRESSION_CONTEXT } from '../constants';
+import { VISUAL_REGRESSION_CONTEXT } from '../shared';
 
 export const getLatestVisualRegressionStatus = async (commitHash: string) => {
   const { data } = await octokit.rest.repos.listCommitStatusesForRef({
