+++ action/run.test.ts
@@ -3,7 +3,7 @@ import { exec } from '@actions/exec';
 import { getInput, getMultilineInput, setFailed } from '@actions/core';
 import { octokit } from './octokit';
 import { sync } from 'glob';
-import { VISUAL_REGRESSION_CONTEXT } from '../constants';
+import { VISUAL_REGRESSION_CONTEXT } from '../shared';
 import { expect } from '@jest/globals';
 
 jest.mock('glob');
