+++ app/frontend/cypress/e2e/home-page.spec.ts
