+++ app/frontend/cypress/fixtures/base-image-update-rejection.json
