+++ app/frontend/cypress/fixtures/new-images-only.json
