+++ app/frontend/cypress/fixtures/no-new-images.json
