+++ app/frontend/cypress/support/component-index.html
