+++ app/frontend/cypress/utils/makeMockAdapter.ts
