--- constants/index.ts
@@ -1,1 +1,0 @@
-export const VISUAL_REGRESSION_CONTEXT = 'Visual Regression';
